==> 00-RAW-Data_Process.ipynb <==
import json
import pandas as pd
import os
import tqdm

# 实体数据

file_list = os.listdir('../entities/')
for one_file in file_list:
    print(one_file)
    with open(f'../entities/{one_file}') as f:
        data = f.read()
    all_data_list = data.split('}\n')
    del data
    all_data_list = [json.loads(one+'}') for one in all_data_list[:-1]]
    one_pkl = one_file.split('.')[0]+'.pkl'
    one_set = pd.DataFrame(all_data_list)
    del all_data_list
    one_set.to_pickle(f'data/entities/{one_pkl}')
    print(one_set.head(2))

one_set = pd.read_pickle(f'data/entities/user.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/course.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/comment.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/concept.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/other.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/paper.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/problem.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/reply.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/video.pkl')
one_set.head()

one_set = pd.read_pickle(f'data/entities/teacher.pkl')
one_set.head()

# 用户问题数据

file_list = os.listdir('../relations/')
print(file_list)
for one_file in file_list:
    if 'json' in one_file:
        print('Processing...',one_file)
        with open(f'../relations/{one_file}') as f:
            data = f.read()
        all_data_list = data.split('}\n')[:-1]
        del data
        if 'problem' in one_file:
            all_user_problem = pd.DataFrame()
            for i in range(14):
                all_need_data = []
                for one in tqdm.tqdm(all_data_list[:10000000]):
                    all_need_data.append(json.loads(one+'}'))
                del all_data_list[:10000000]
                user_problem = pd.DataFrame(all_need_data)
                del all_need_data
                user_problem['user_id'] = user_problem['user_id'].apply(lambda x: int(x.split('_')[1]))
                user_problem['problem_id'] = user_problem['problem_id'].apply(lambda x: int(x.split('_')[1]))
                all_user_problem = pd.concat([all_user_problem, user_problem], axis=0)
            pd.to_pickle(all_user_problem,'data/relations/all_user_problem.pkl')
            del all_user_problem
        else:
            all_need_data = []
            for one in tqdm.tqdm(all_data_list):
                all_need_data.append(json.loads(one+'}'))
            del all_data_list
            one_pkl = one_file.split('.')[0]+'.pkl'
            one_set = pd.DataFrame(all_need_data)
            del all_need_data
            one_set.to_pickle(f'data/relations/{one_pkl}')
            del one_set

# 获取关系数据里面的txt

## 获取视频的ccid数据

with open(f'../relations/video_id-ccid.txt') as f:
    data = f.read()
all_data_list = data.split('\n')[:-1]

## 获取练习和问题的关系数据

with open(f'../relations/exercise-problem.txt') as f:
    data = f.read()
all_data_list = data.split('\n')[:-1]
all_data = []
for one in all_data_list:
    v_c = one.split('\t')
    one_e_id = int(v_c[0].split('_')[1])
    one_q_id = int(v_c[1].split('_')[1])
    all_data.append({'e_id':one_e_id,'q_id':one_q_id})
all_data = pd.DataFrame(all_data)
pd.to_pickle(all_data,'data/relations/exercise_problem.pkl')
all_data.head()

## 获取概念与问题的关系

with open(f'../relations/concept-problem.txt') as f:
    data = f.read()
data = data.split('\n')[:-1]
concept_problem = []
for one in tqdm.tqdm(data):
    v_c = one.split('\t')
    one_c = v_c[0].split('_')[1]
    one_s = v_c[0].split('_')[2]
    one_q_id = int(v_c[1].split('_')[1]) 
    concept_problem.append({'concept':one_c,'subject':one_s,'q_id':one_q_id})
concept_problem = pd.DataFrame(concept_problem)
pd.to_pickle(concept_problem,'data/relations/concept_problem.pkl')
concept_problem.head()

## 获取用户评论的数据

with open(f'../relations/user-comment.txt') as f:
    data = f.read()
data = data.split('\n')[:-1]
data[0]
concept_problem = []
for one in tqdm.tqdm(data):
    v_c = one.split('\t')
    one_c = v_c[0].split('_')[1]
    one_s = v_c[0].split('_')[1]
    concept_problem.append({'stu_id':int(one_c),'comment_id':int(one_s)})
concept_problem = pd.DataFrame(concept_problem)
pd.to_pickle(concept_problem,'data/relations/user_comment.pkl')
concept_problem.head()

## 获取评论与概念的关系

with open(f'../relations/concept-comment.txt') as f:
    data = f.read()
data = data.split('\n')[:-1]
concept_problem = []
for one in tqdm.tqdm(data):
    v_c = one.split('\t')
    one_c = v_c[0].split('_')[1]
    one_s = v_c[0].split('_')[2]
    one_q_id = int(v_c[1].split('_')[1]) 
    concept_problem.append({'concept':one_c,'subject':one_s,'comment':one_q_id})
concept_problem = pd.DataFrame(concept_problem)
pd.to_pickle(concept_problem,'data/relations/concept_comment.pkl')
concept_problem.head()

## 获取概念与视频的关系

with open(f'../relations/concept-video.txt') as f:
    data = f.read()
data = data.split('\n')[:-1]
concept_problem = []
for one in tqdm.tqdm(data):
    v_c = one.split('\t')
    one_c = v_c[0].split('_')[1]
    one_s = v_c[0].split('_')[2]
    concept_problem.append({'concept':one_c,'subject':one_s,'v_ccid':v_c[1]})
concept_problem = pd.DataFrame(concept_problem)
pd.to_pickle(concept_problem,'data/relations/concept_video.pkl')
concept_problem.head()

## 获取概念与论文的关系

with open(f'../relations/concept-paper.txt') as f:
    data = f.read()
data = data.split('\n')[:-1]
concept_problem = []
for one in tqdm.tqdm(data):
    try:
        v_c = one.split('\t')
        one_c = v_c[0].split('_')[1]
        one_s = v_c[0].split('_')[2]
    #     one_q_id = v_c[1].split('_')[1]
        concept_problem.append({'concept':one_c,'subject':one_s,'paper_ccid':v_c[1]})
    except:
        continue
concept_problem = pd.DataFrame(concept_problem)
pd.to_pickle(concept_problem,'data/relations/concept_paper.pkl')
concept_problem.head()

# 处理prerequisites里的概念的先后关系

with open(f'../prerequisites/math.json') as f:
    data = f.read()
data = data.split('}\n')
data = [json.loads(one+'}') for one in data[:-1]]
data = pd.DataFrame(data)
pd.to_pickle(data,'data/relations/math.pkl')
data.head()

with open(f'../prerequisites/cs.json') as f:
    data = f.read()
data = data.split('}\n')
data = [json.loads(one+'}') for one in data[:-1]]
data = pd.DataFrame(data)
pd.to_pickle(data,'data/relations/cs.pkl')
data.head()

with open(f'../prerequisites/psy.json') as f:
    data = f.read()
data = data.split('}\n')
data = [json.loads(one+'}') for one in data[:-1]]
data = pd.DataFrame(data)
pd.to_pickle(data,'data/relations/psy.pkl')
data.head()
